Move splash to new component
