+++ pages/index.js
@@ -1,25 +1,12 @@
 import styles from "./Home.module.css"
-import LogoNoBg from "@/assets/logo_nobg.svg";
-
-const SplashLogo = () => {
-  return <div className={styles.splash_logo_bar}>
-    <LogoNoBg className={styles.splash_logo}></LogoNoBg>
-    <div>
-      <button>Order now!</button>
-    </div>
-  </div>
-}
-
-const SplashImage = () => {
-  return <div className={styles.splash}>
-    <SplashLogo></SplashLogo>
-  </div>
-}
+import SplashImage from "@/components/home/splash/Splash";
 
 const Home = () => {
   return <div className={styles.home}>
-    <SplashImage>
-    </SplashImage>
+    <SplashImage></SplashImage>
+    <div className={styles.section_header}>
+      <div><h1>Hello</h1></div>
+    </div>
   </div>
 }
